+++ src/main/java/com/furiosaming/clusterAnalysis/model/Image.java
@@ -11,14 +11,14 @@ import java.util.Map;
 public class  Image implements Serializable {
 
     private int id;
-    private Map<String, Integer> characteristics;
+    private Map<String, Double> characteristics;
 
     public Image(int id){
         this.id = id;
         this.characteristics = new HashMap<>();
     }
 
-    public void addCharacteristics(String key, Integer value){
+    public void addCharacteristics(String key, Double value){
         characteristics.put(key, value);
     }
